fix: Resolve circular dependency issue
